requirements.txt
dill
dnspython
evidently
fastapi
httptools
imblearn
mypy-boto3-s3
pip-chill
pymongo
python-dotenv
types-s3transfer
uvicorn
watchfiles
websockets
wincertstore
xgboost
neuro-mf
-e .